requirements.txt
pytest==7.1.2
selenium==3.141.0
webdriver_manager==3.5.2
pytest-html==3.1.1
paramiko==2.10.4
PyMySQL==1.0.2
sshtunnel==0.4.0
allure-pytest==2.9.45
pluggy==1.0.0
py==1.11.0
allure-python-commons==2.8.31
decorator==4.4.2
httplib2==0.18.1
json2html==1.3.0
jsonpath-ng==1.5.3
matplotlib==3.1.2
nose==1.3.7
numpy==1.17.4
numpydoc==0.7.0
pandas==0.25.3
pytest==6.2.2
pytest-forked==1.3.0
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.0
requests==2.26.0
unittest2==1.1.0
pytest-check==1.0.4
requests==2.26.0
